Song management + picker in processor

+++ app/src/main/java/dev/leonardini/rehcorder/db/SongRecording.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "song_recording")
 data class SongRecording(
-    @PrimaryKey(autoGenerate = true) val uid: Int,
-    @ColumnInfo(name = "song_id") val songId: Int,
-    @ColumnInfo(name = "recording_id") val recordingId: Int,
+    @PrimaryKey(autoGenerate = true) val uid: Long,
+    @ColumnInfo(name = "song_id") val songId: Long,
+    @ColumnInfo(name = "recording_id") val recordingId: Long,
 )
